+++ tests/unit/algorithms/segmentation/adapters/mmseg/datasets/pipelines/__init__.py
@@ -1,0 +1,4 @@
+"""Test for otx.algorithms.segmentation.adapters.mmseg.datasets.pipelines"""
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
