+++ tests/unit/algorithms/segmentation/adapters/mmseg/datasets/pipelines/test_compose.py
