+++ tests/unit/algorithms/segmentation/adapters/mmseg/models/backbones/__init__.py
@@ -1,0 +1,4 @@
+"""Test for otx.algorithms.segmentation.adapters.mmseg.models.backbones."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
