+++ tests/unit/algorithms/segmentation/adapters/mmseg/models/heads/__init__.py
@@ -1,4 +1,4 @@
-"""Test for otx.mpa.modules.models.backbones."""
+"""Test for otx.algorithms.segmentation.adapters.mmseg.models.heads."""
 
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
