+++ tests/unit/algorithms/segmentation/adapters/mmseg/models/scalar_schedulers/__init__.py
@@ -1,4 +1,4 @@
-"""Test for otx.mpa.modules.models.scheculers"""
+"""Test for otx.algorithms.segmentation.adapters.mmseg.models.scheculers"""
 
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
