+++ tests/unit/algorithms/segmentation/adapters/mmseg/models/scalar_schedulers/test_schedulers.py
