+++ tests/unit/algorithms/segmentation/adapters/mmseg/utils/__init__.py
@@ -1,4 +1,4 @@
-"""Test for otx.algorithms.segmentation.adapters.mmseg.data"""
+"""Test for otx.algorithms.segmentation.adapters.mmseg.utils."""
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
